MAJ après deploy render

+++ db/migrate/20250529130005_remove_availability_slots_from_artisans.rb
@@ -1,5 +1,8 @@
 class RemoveAvailabilitySlotsFromArtisans < ActiveRecord::Migration[7.1]
   def change
-    remove_column :artisans, :availability_slots, :jsonb
+    # On vérifie si la colonne existe avant de la supprimer
+    if column_exists?(:artisans, :availability_slots)
+      remove_column :artisans, :availability_slots, :jsonb
+    end
   end
 end
